api/requirements.txt
flask
paho.mqtt
pyopenssl